scripts/requirements.txt
python-amazon-paapi==5.0.5

